Add network mapper code and test
